+++ backend/src/main/java/com/kongdak/global/security/jwt/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package com.kongdak.config;
+package com.kongdak.global.security.jwt;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
